backend/requirements.txt
spotipy
python-dotenv
bandcamp-search
fastapi
itsdangerous
uvicorn[standard]
sse-starlette
requests
beautifulsoup4
soundcloud-v2
